fix: 필터링 헤더가 홈의 모든 탭에서 나타나는 오류 해결
